pyproject.toml
[project]
name = "spraycoater-v3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "printrun>=2.1.0",
    "pyserial>=3.5",
]
